Refactor relayer package
